+++ packages/react-instantsearch-dom-geo/src/elements/__tests__/createHTMLMarker.js
