+++ packages/react-instantsearch-dom-geo/src/elements/createHTMLMarker.js
